src/requirements.txt
Django>=3.0.6,<3.1
psycopg2>=2.8.5,<2.9
Pillow
gunicorn>=20.0.4,<20.1
django-allauth>=0.41.0,<0.42
WeasyPrint>=51,<=59
django-crispy-forms>=1.9.0,<1.10
django-filter>=2.2.0,<2.3
django-tables2>=2.3.1,<2.4
djangorestframework>=3.11.0,<4
tablib>=1.1.0,<1.5.0
openpyxl>=3.0.3,<3.1
jdcal>=1.4.1,<1.5
et-xmlfile>=1.0.1,<1.1
django-taggit==1.2.0
django-summernote==0.8.11.6
django-vote==2.1.7
django-notifications-hq>=1.6.0,<1.7
django-admin-list-filter-dropdown==1.0.3
requests>=2.25.1
python-docx==0.8.10